Remove superfluous package.
