+++ src/rhsmlib/dbus/services/base_properties.py
@@ -2,7 +2,7 @@ import logging
 
 import dbus
 
-from rhsmlib.dbus.common import dbus_utils
+from rhsmlib.dbus import dbus_utils
 
 log = logging.getLogger(__name__)
