+++ src/rhsmlib/dbus/services/examples/examples.py
@@ -1,7 +1,7 @@
 import dbus.service
 import slip.dbus
 
-import rhsmlib.dbus.common as common
+import rhsmlib.dbus as common
 
 from rhsmlib.dbus.services import base_service
 from rhsmlib.dbus.services import base_properties
